Change CreateMinimal to return Ok with product object
Modified the CreateMinimal method in ProductController to return a 200 OK response with the product data instead of a 201 Created response with the URI of the newly created product. This change simplifies the response and may align with a different API design choice.

Change CreateMinimal to return Ok with product object

The return type of the `CreateMinimal` method in the `ProductController` class has been changed. Previously, it returned a `Created` response with the URI of the newly created product. Now, it returns an `Ok` response with the product object. This change simplifies the response and aligns with a different API design choice.