Add GetByProductIdAsync method to ManufacturerService
A new method GetByProductIdAsync has been added to the
IManufacturerService interface to retrieve a ManufacturerDto
based on a given product ID. The ManufacturerService class
has been updated to implement this method, which queries
the ProductManufacturerMappings table. If no manufacturer
is found, a NotFoundException is thrown. If a manufacturer
is found, it is mapped to a ManufacturerDto and returned.